BJS2-52766-Добавить тест на проверку того, что Spring Context успешно поднимается Немного порефакторил, т.к. с общими изменениями в разных директориях SkillService'а было 2
